+++ config/runtime.exs
@@ -20,6 +20,20 @@ if System.get_env("PHX_SERVER") do
   config :td_qx, TdQxWeb.Endpoint, server: true
 end
 
+config :td_cluster,
+  scope: :truedat,
+  topologies: [
+    truedat: [
+      strategy: TdCluster.Strategy,
+      config: [
+        node_template: System.get_env("RELEASE_NODE_TEMPLATE", "{{service}}@{{hostname}}"),
+        services: [:dd],
+        groups: [:qx],
+        polling_interval: 10_000
+      ]
+    ]
+  ]
+
 if config_env() == :prod do
   maybe_ipv6 = if System.get_env("ECTO_IPV6") in ~w(true 1), do: [:inet6], else: []
 
@@ -33,18 +47,4 @@ if config_env() == :prod do
     timeout: System.get_env("DB_TIMEOUT_MILLIS", "15000") |> String.to_integer(),
     ssl: System.get_env("DB_SSL", "") |> String.downcase() == "true",
     socket_options: maybe_ipv6
-
-  config :libcluster,
-    topologies: [
-      k8s: [
-        strategy: Elixir.Cluster.Strategy.Kubernetes,
-        config: [
-          mode: :ip,
-          kubernetes_node_basename: System.get_env("K8S_NODE_BASENAME", "truedat"),
-          kubernetes_selector: System.get_env("K8S_SELECTOR", ""),
-          kubernetes_namespace: System.get_env("K8S_NAMESPACE", "default"),
-          polling_interval: 10_000
-        ]
-      ]
-    ]
 end