+++ lib/td_qx_web/router.ex
@@ -10,4 +10,10 @@ defmodule TdQxWeb.Router do
 
     get "/ping", PingController, :ping
   end
+
+  scope "/api", TdQxWeb do
+    pipe_through :api
+
+    resources "/data_sets", DataSetController, except: [:new, :edit]
+  end
 end