+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [5.10.0] 2023-07-06
+
+### Added
+
+- [TD-5808] API of enriched DataSets with data structures brought from the dd service
+
+### Changed
+
+- [TD-5912] `.gitlab-ci.yml` adaptations for develop and main branches
+
 ## [5.8.0] 2023-06-05
 
 ### Added